Round-11 review fixes: file split, late-attachment editor clear
Each stateful todo/checklist control moves to its own file per the
one-major-type-per-file standard (section, summary line, item line, add
row, attachment button, field support, popover presenter, status glyph,
compact status line), and the checklist fields clear the field editor's
background after the deferred focus grab too — cells configured before
window attachment previously restored the dark editor box.
